% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxid2Name.R
\name{locate_taxonomy_cache}
\alias{locate_taxonomy_cache}
\title{Locate taxonomic cache}
\usage{
locate_taxonomy_cache()
}
\description{
Find where \code{\link[=db_download_ncbi]{db_download_ncbi()}} downloads the taxonomy database to
}
