% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxid2Name.R
\name{taxid2lineage}
\alias{taxid2lineage}
\title{Taxid to lineage}
\usage{
taxid2lineage(
  taxids,
  ranks_to_include = c("no rank", "superkingdom", "phylum", "class", "order", "family",
    "genus", "species", "strain"),
  show_ranks = FALSE,
  ultimate_ancestor = NULL,
  special_taxid_names = c(unclassified = -1)
)
}
\arguments{
\item{taxids}{ncbi taxonomy ids (numeric)}

\item{ranks_to_include}{which ranks to include in lineage string (character)}

\item{show_ranks}{Should ranks of each lineage level be added to (boolean)}

\item{ultimate_ancestor}{Text to add to (string)}

\item{special_taxid_names}{for taxids <= 0 you can define special names. (named numeric vector where names = names and values = taxids)}
}
\value{
'>' delimited lineage string for each taxid supplied (character)
}
\description{
Take an ncbi taxonomy id (e.g. 562 - E. coli) and return a taxanomic lineage string
}
\examples{
\dontrun{
taxid2lineage(562, ranks_to_include=c("superkingdom","family", "genus", "species"))
# Bacteria>Enterobacteriaceae>Escherichia>Escherichia coli
}
}
