% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxid2Name.R
\name{taxid2name}
\alias{taxid2name}
\title{NCBI Taxid to Scientific Name}
\usage{
taxid2name(taxids, special_taxid_names = c(unclassified = -1))
}
\arguments{
\item{taxids}{vector of taxids (numeric)}

\item{special_taxid_names}{give custom names for taxids <=0. Supply numeric vector where values correspond to taxids <=0 and names correspond to the custom names we want to use for each of these 'taxids'.}
}
\value{
scientific names for each taxid, or NA if no taxanomic match could be found. for negative taxids, value of 'negative_taxid_name' will dictate the name returned
}
\description{
Faster taxid2name for long vectors with repeated taxids.
You may need to first run \code{taxizedb::db_download_ncbi()}
}
